create interface
