+++ src/components/IndoorNavigation.tsx
@@ -5,30 +5,7 @@ import { useIndoor } from "../data/IndoorContext";
 import { useCoords } from "../data/CoordsContext";
 import { IndoorFeatureCollection } from '../interfaces/IndoorFeature';
 import { FeatureCollection, LineString, Position } from 'geojson';
-
-// Starting points types
-export enum EntryPointType {
-  ELEVATOR = 'elevator',
-  ESCALATOR = 'escalator',
-  STAIRS = 'stairs',
-  ANY = 'any' // Default - will try all types
-}
-
-// Define types for our graph
-interface GraphNode {
-  id: string;
-  position: Position; // [longitude, latitude]
-  neighbors: string[];
-  isEntryPoint?: boolean;
-  entryPointType?: EntryPointType;
-  isRoom?: boolean;
-  roomNumber?: string;
-  isCorridor?: boolean;
-}
-
-interface Graph {
-  nodes: { [id: string]: GraphNode };
-}
+import { EntryPointType, GraphNode, Graph } from '../interfaces/IndoorGraph';
 
 // A* pathfinding algorithm
 const findPath = (
